// snl-gms-mock/packages/common-model/src/ts/user-profile/types.ts
export const AllHideEmptyWaveformRows = [
  /** if the empty waveform rows are always hidden */
  'ALWAYS_ON',
  /** if the empty waveform rows are never hidden */
  'ALWAYS_OFF',
  /** if whether the empty waveform rows are hidden is based on WaveformMode */
  'PER_MODE'
] as const;

export type HideEmptyWaveformRows = (typeof AllHideEmptyWaveformRows)[number];

export const AllNumberVisibleWaveformsOptions = ['PER_MODE', 'CUSTOM'] as const;
export type NumberVisibleWaveformsOption = (typeof AllNumberVisibleWaveformsOptions)[number];

export interface GlobalPreferences {
  readonly colorMap: ColorMapName;
  readonly currentTheme: UITheme['name'];
  readonly hideEmptyWaveformRows: HideEmptyWaveformRows;
  readonly numberVisibleWaveformsOption: NumberVisibleWaveformsOption;
  readonly customNumberVisibleWaveforms: number;
}

type GlobalPreferencesKey = keyof GlobalPreferences;
export type GlobalPreferencesProperty = GlobalPreferences[GlobalPreferencesKey];

export interface WorkspaceLayout {
  readonly name: string;
  readonly layoutConfiguration: string;
}
// ...
export interface MapDisplaySettings {
  readonly twoDimensional: boolean;
  readonly layerVisibility: Record<string, boolean>;
  readonly syncToWaveformDisplayVisibleTimeRange: boolean;
}
// ...
export interface TabularDisplayLayout {
  /** Record of `{ columnId: isVisible }` */
  readonly columnVisibility: Record<string, boolean>;
  /** Record of `{ columnId: 'left' | 'right' }` */
  readonly pinnedColumns: Record<string, 'left' | 'right'>;
  readonly columnOrder: string[];
  /** Record of `{ columnId: columnWidthPixels }` */
  readonly columnWidths: Record<string, number>;
}

export type TabularDisplayLayoutKey = keyof TabularDisplayLayout;

export type TabularDisplayColumnWidths = {
  colId: string;
  width: number;
}[];

export type TabularDisplayLayoutProperty =
  | TabularDisplayLayout[Exclude<TabularDisplayLayoutKey, 'columnWidths'>]
  | TabularDisplayColumnWidths;

export interface DualDisplayLayout {
  displayOne: TabularDisplayLayout;
  displayTwo: TabularDisplayLayout;
}

type DualDisplayLayoutKey = keyof DualDisplayLayout;

export type TableDisplayNamePath = [
  displayKey: IanDisplays,
  ...dualDisplayKeys: DualDisplayLayoutKey[]
];

export interface SignalDetectionDisplayLayout extends TabularDisplayLayout {
  /** Record of `{ signalDetectionType: isVisible }` */
  readonly visibleSignalDetections: Record<string, boolean>;
}

export interface EventDisplayLayout extends TabularDisplayLayout {
  /** Record of `{ eventType: isVisible }` */
  readonly visibleEvents: Record<string, boolean>;
}

export interface AzimuthSlownessDisplayLayout extends Pick<TabularDisplayLayout, 'columnOrder'> {
  readonly thumbnailSize: FkThumbnailSize;
  readonly thumbnailColumnWidth: number;
  /** Denotes if FK step lines should be displayed in beams/traces panel */
  readonly fkStepsVisible: boolean;
  /** Denotes if frequency preview thumbnails should be displayed */
  readonly fkPreviewsVisible: boolean;
  /** Denotes if beams and traces panel is visible */
  readonly fkBeamsAndTracesVisible: boolean;
}

export interface LocationDisplayLayout extends Pick<DualDisplayLayout, 'displayOne'> {
  displayTwo: Omit<TabularDisplayLayout, 'pinnedColumns'>;
}

export interface MagnitudeDisplayLayout extends DualDisplayLayout {
  /** Record of `{ magnitudeType: isVisible }` */
  visibleMagnitudeTypes: Record<string, boolean>;
}

export type AzimuthSlownessPreferencesKey = Exclude<
  keyof AzimuthSlownessDisplayLayout,
  TabularDisplayLayout
>;
export type AzimuthSlownessPreferencesProperty =
  AzimuthSlownessDisplayLayout[AzimuthSlownessPreferencesKey];

export interface WaveformDisplayLayout {
  readonly visibleSignalDetections: Record<string, boolean>;
}

export interface StationVisibilityList {
  readonly name: string;
  readonly stations: EntityReference<'name', Station>[];
}

/**
 * User UserLayout COI
 *
 * A URI encoded json representation of a golden layout.
 * Includes metadata like supported user interface modes and the name of the layout.
 */
export interface UserLayout {
  readonly name: string;
  readonly layoutConfiguration: string;
}

/**  User profile COI */
export interface UserProfile {
  readonly userId: string;
  readonly userPreferences: UiUserPreferences;
}

export interface DisplayLayouts {
  [IanDisplays.AZIMUTH_SLOWNESS]: AzimuthSlownessDisplayLayout;
  [IanDisplays.EVENTS]: EventDisplayLayout;
  [IanDisplays.SIGNAL_DETECTIONS]: SignalDetectionDisplayLayout;
  [IanDisplays.LOCATION]: LocationDisplayLayout;
  [IanDisplays.STATION_PROPERTIES]: DualDisplayLayout;
  [IanDisplays.MAGNITUDE]: MagnitudeDisplayLayout;
}

/**
 * KeyboardShortcuts collection specific to userPreferences.
 * Each shortcut type maps to a keyboardShortcut name and a list of additional combo strings
 *
 * @example
 * {
 *   clickEvents: {
 *     associateSelectedSignalDetections: ["shift+0"]
 *   }
 * }
 */
export type UserPreferencesKeyboardShortcuts = Modify<
  KeyboardShortcuts,
  {
    clickEvents: Partial<Record<ClickEventDefinitions, string[]>>;
    middleClickEvents: Partial<Record<never, string[]>>;
    rightClickEvents: Partial<Record<never, string[]>>;
    doubleClickEvents: Partial<Record<DoubleClickDefinitions, string[]>>;
    dragEvents: Partial<Record<DragEventDefinitions, string[]>>;
    scrollEvents: Partial<Record<ScrollEventDefinitions, string[]>>;
    hotkeys: Partial<Record<HotkeyDefinitions, string[]>>;
  }
>;

/**
 * Holds info about choices the choices a user has made about displays in the UI
 */
export interface UiUserPreferences {
  readonly defaultWorkspaceLayout: string;
  readonly globalPreferences: GlobalPreferences;
  readonly workspaceLayouts: WorkspaceLayout[];
  readonly customFilterList?: FilterList;
  readonly mapDisplaySettings: MapDisplaySettings;
  readonly displayLayouts: DisplayLayouts;
  readonly waveformDisplayLayout: WaveformDisplayLayout;
  readonly stationVisibilityLists: StationVisibilityList[];
  readonly keyboardShortcuts: UserPreferencesKeyboardShortcuts;
}